Added a castle sprite for the player's castle
-Added Hitbox.cs and Projectile.cs
-Edited GetType in Action.cs to be GetActionType so it does not hide Object.GetType()

+++ Defend Your Castle/Defend Your Castle/Enemies/EnemyAI/MoveForward.cs
@@ -12,9 +12,9 @@ namespace Defend_Your_Castle
     public sealed class MoveForward : Action
     {
         //The X to stop
-        public float XStop;
+        public int XStop;
 
-        public MoveForward(Enemy enem, Animation anim, float xstop) : base(enem, anim, ActionType.Moving)
+        public MoveForward(Enemy enem, Animation anim, int xstop) : base(enem, anim, ActionType.Moving)
         {
             XStop = xstop;
         }
